+++ src/api/controllers/instance.controller.ts
@@ -12,6 +12,7 @@ import { RabbitmqService } from '../integrations/rabbitmq/services/rabbitmq.serv
 import { SqsService } from '../integrations/sqs/services/sqs.service';
 import { TypebotService } from '../integrations/typebot/services/typebot.service';
 import { WebsocketService } from '../integrations/websocket/services/websocket.service';
+import { ProviderFiles } from '../provider/sessions';
 import { RepositoryBroker } from '../repository/repository.manager';
 import { AuthService, OldToken } from '../services/auth.service';
 import { CacheService } from '../services/cache.service';
@@ -43,6 +44,7 @@ export class InstanceController {
     private readonly cache: CacheService,
     private readonly chatwootCache: CacheService,
     private readonly messagesLostCache: CacheService,
+    private readonly providerFiles: ProviderFiles,
   ) {}
 
   private readonly logger = new Logger(InstanceController.name);
@@ -111,6 +113,7 @@ export class InstanceController {
           this.cache,
           this.chatwootCache,
           this.messagesLostCache,
+          this.providerFiles,
         );
       } else {
         instance = new BaileysStartupService(
@@ -120,6 +123,7 @@ export class InstanceController {
           this.cache,
           this.chatwootCache,
           this.messagesLostCache,
+          this.providerFiles,
         );
       }