+++ src/api/services/channels/whatsapp.business.service.ts
@@ -23,6 +23,7 @@ import {
   SendTextDto,
 } from '../../dto/sendMessage.dto';
 import { ContactRaw, MessageRaw, MessageUpdateRaw, SettingsRaw } from '../../models';
+import { ProviderFiles } from '../../provider/sessions';
 import { RepositoryBroker } from '../../repository/repository.manager';
 import { Events, wa } from '../../types/wa.types';
 import { CacheService } from './../cache.service';
@@ -36,6 +37,7 @@ export class BusinessStartupService extends ChannelStartupService {
     public readonly cache: CacheService,
     public readonly chatwootCache: CacheService,
     public readonly messagesLostCache: CacheService,
+    private readonly providerFiles: ProviderFiles,
   ) {
     super(configService, eventEmitter, repository, chatwootCache);
     this.logger.verbose('BusinessStartupService initialized');