+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Feature
 
 * Now in the manager, when logging in with the client's apikey, the listing only shows the instance corresponding to the provided apikey (only with MongoDB)
+* New global mode for rabbitmq events
 
 # 1.7.5 (2024-05-21 08:50)